+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed help generation to a prettier one
+- Changed weightage for the meme gay command
 
 ## [0.4.0] - 2024-10-22