+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.5.0 - 2024-11-15
+
 ### Added
 
 - Introduction of database to track persistent data.
@@ -19,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed help generation to a prettier one
 - Changed weightage for the meme gay command
+
+### Removed
+
+- Removed external youtube oauth plugin
 
 ## [0.4.0] - 2024-10-22